Add files via upload
This simulation shows the average value of a cribbage hand is 8 pts.  It randomly deals out six cards, and chooses the highest scoring card hand (assuming no points from the starter).  Then a starter is added and the point value of the hand is calculated.  Currently doesn't score flushes.  One other minor bug (only 14 of the 15 possible 6C4 hand combinations are tested).

Plan to use this to make a game with an AI cribbage opponent.